♻️  [Refactor]: Use Logo Component & Rename File
- Logo Component 사용
- isPathMatched -> isRoutePathMatched로 일관성 있게 작성
