+++ src/components/organisms/header/Logo.tsx
@@ -3,15 +3,21 @@ import { Link } from 'react-router-dom';
 
 import Icon from '@/components/atoms/Icon';
 import cn from '@/libs/cn';
+import { routePaths } from '@/constants/route';
 
 type LogoProps = ComponentProps<typeof Link> & {
   linkStyle?: string;
   iconStyle?: string;
 };
 
-export default function Logo({ linkStyle, iconStyle, ...others }: LogoProps) {
+export default function Logo({
+  linkStyle,
+  iconStyle,
+  to,
+  ...others
+}: LogoProps) {
   return (
-    <Link className={linkStyle} {...others}>
+    <Link to={to || routePaths.root} className={linkStyle} {...others}>
       <Icon type="logo" className={cn('h-[3rem] w-[5.9375rem]', iconStyle)} />
     </Link>
   );
